Generate root inode dumps in CI and drop a redundant cast
The inode cross-validation test requires .inodedump fixtures and fails
outright when none exist, so an unvalidated parser cannot pass by having
nothing to compare against. CI was building superblock dumps only, which
would have tripped that guard. It now dumps the root inode for every
geometry it builds, and removes the dump alongside the image when a
geometry is skipped.

+++ src/inode.rs
@@ -307,7 +307,7 @@ impl Inode {
             mode: be16(buf, 2),
             version,
             format: Format::from_raw(buf[5])?,
-            aformat: Format::from_raw(buf[83] as u8)?,
+            aformat: Format::from_raw(buf[83])?,
             uid: be32(buf, 8),
             gid: be32(buf, 12),
             nlink: be32(buf, 16),
